v0.5
v0.5 introduces Extras, the other half of the framework, which includes
pre-made goodies like buttons and graphs (!), all made from CSS and
ready for you to use in your next web project! This part of the
framework is not complete, and will certainly grow before the v1.0
release. Otherwise, the “website” folder has been renamed “template”
for clarity and minor tweaks were made to the template CSS files.