v0.6
As I struggle to find an identity for the Jatsby Web Framework, it’s
been stripped of all but the template, which has been consolidated into
one html file and one css file. I’m hoping to add “extras” (buttons,
etc.) eventually, but since my taste changes so often, that will be
difficult to do. As it is, it’s not bad, and I still use it as a
template for all my web projects. Unfortunately, it doesn’t do much to
stand out from other frameworks.